Name refactoring.
