+++ benchmarks/fs2/src/main/scala/laserdisc/fs2/parallel/ParallelLoadLaserdiscCatsByteBench.scala
@@ -8,7 +8,7 @@ import cats.effect.{Blocker, ContextShift, IO, Timer}
 import cats.syntax.flatMap._
 import laserdisc.auto._
 import laserdisc.fs2.parallel.SetUpLaserdiscCatsByte.LaserdiscCatsByteSetUp
-import laserdisc.fs2.parallel.baseline.BenchRedisBytesChannel
+import laserdisc.fs2.parallel.channels.ByteInBitVectorOutChannel
 import laserdisc.fs2.parallel.testcases.TestCasesLaserdiscByte
 import log.effect.fs2.SyncLogWriter
 import log.effect.{LogLevels, LogWriter}
@@ -51,7 +51,7 @@ object SetUpLaserdiscCatsByte {
 
     val resource = Blocker[IO] evalMap { bl =>
       RedisAddress("localhost", 6379).toInetSocketAddress[IO] map { address =>
-        BenchRedisBytesChannel[IO](address, writeTimeout = Some(10.seconds), readMaxBytes = 8 * 1024 * 1024)(bl)
+        ByteInBitVectorOutChannel[IO](address, writeTimeout = Some(10.seconds), readMaxBytes = 8 * 1024 * 1024)(bl)
       }
     }
